Move client side code out of utilites
