+++ src/components/uploadPage/uploadPage.tsx
@@ -8,8 +8,8 @@ import DeployFormSummary from "./sections/deployFormSummary";
 import AlreadyExists from "./sections/alreadyExists";
 import ConfirmOverride from "./sections/confirmOverride";
 import InvalidSettings from "./sections/invalidSettings";
-import { uploadAndInstallFile, validateSelectedInstrumentExists, checkInstrumentSettings } from "./uploadProcess";
-import { roundUp } from "../../utilities";
+import { uploadAndInstallFile, validateSelectedInstrumentExists, checkInstrumentSettings } from "../../client/componentProcesses/uploadProcess";
+import { roundUp } from "../../utilities/maths";
 import Breadcrumbs from "../breadcrumbs";
 import { activateInstrument, deleteInstrument } from "../../client/instruments";
 import { InstrumentSettings, Instrument } from "blaise-api-node-client";
