+++ src/utilities/textFormatting/textFormatting.test.ts
@@ -1,0 +1,7 @@
+import { formatText } from "./textFormatting";
+
+describe("formatText", () => {
+  it("formats text to title case without underscores", () => {
+    expect(formatText("this_is_a_sentance")).toEqual("This is a sentance");
+  });
+});
